feat(strmangle): Pulling strmangle implementation into struct with interface
Designed to facilitate programmatic overrides to strmangle implementation by setting DefaultMangler
